Activate eslint for vulcan-users and make it pass
